Module egonlang_core::ast

Structs

ExprAssign

Expression for assigning a value expression to an identifier

ExprBlock

Expression creating a lexical scope that optionally returns a value

ExprFn

Expression creating an anonymous function

ExprIdentifier

Expression representing literal values e.g. string, number, bool

ExprIf

Expression returning a value based on a condition

ExprInfix

Expression preforming an infix operation

ExprList

Expression creating a list of same typed values

ExprPrefix

Expression preforming a prefix operation

ExprRange

Expression creating a range of values

ExprTuple

Expression creating a fixed sized collection of mixed typed values

ExprType

Expression representing a type

Identifier

Module

A collection of Stmt representing an Egon code file.

StmtAssertType

A statement to assert an expression’s type

StmtAssign

A statement that declares and/or initalizes a variable or constant

StmtExpr

A statement that evaluates an expression then consumes the value.

StmtFn

A statement declaring a function

StmtReturn

A statement to return a value from a function

StmtTypeAlias

A statement that declares a type alias

TypeRef

Value representing a type

Enums

AstNode

AST nodes contained in the Module

Expr

Expressions

ExprLiteral

Expression representing literal values e.g. string, number, bool

OpInfix

OpPrefix

Stmt

A statement (e.g. that does not produce a value)

Type Aliases

ExprS

A tuple containing an expression and it’s span e.g. (expr, span)

StmtS

A tuple containing a statement and it’s span e.g. (stmt, span)