Enum egonlang_core::ast::Expr

pub enum Expr {

    Unit,
    Literal(ExprLiteral),
    Identifier(ExprIdentifier),
    Block(Box<ExprBlock>),
    List(Box<ExprList>),
    Tuple(Box<ExprTuple>),
    Infix(Box<ExprInfix>),
    Prefix(Box<ExprPrefix>),
    Assign(Box<ExprAssign>),
    If(Box<ExprIf>),
    Fn(Box<ExprFn>),
    Range(ExprRange),
    Type(ExprType),
    Call(Box<ExprCall>),

}

Expressions

Variants

Unit

Expression returning no value

e.g. void, () in other languages

();

Literal(ExprLiteral)

Expression representing literal values e.g. string, number, bool

123;
"example";
false;

Identifier(ExprIdentifier)

Expression representing an identifier

let a: number = 123; // `a` & `number` are identifier expressions
a + 5; // `a` is an identifier expression

Block(Box<ExprBlock>)

Expression creating a lexical scope that optionally returns a value

let a: () = { 123; };
let b: number = { 456 };

List(Box<ExprList>)

Expression creating a list of same typed values

[];
[1, 2, 3];

Tuple(Box<ExprTuple>)

Expression creating a fixed sized collection of mixed typed values

(1,);
(1, 2, 3,);

Infix(Box<ExprInfix>)

Expression preforming an infix operation

1 + 2;
true != false;

Prefix(Box<ExprPrefix>)

Expression preforming a prefix operation

-10;
!false;

Assign(Box<ExprAssign>)

Expression for assigning a value expression to an identifier

let a: number;
a = 123;

This is not allowed for constants

const a: number = 123;
a = 1; // This will generate an error

If(Box<ExprIf>)

Expression returning a value based on a condition

let a: number = if (true) { 123 } else { 456 };

Fn(Box<ExprFn>)

Expression creating an anonymous function

(a: number): number => { a + 10 };

Range(ExprRange)

Expression creating a range of values

0..10;
10..200;
100..0;

Type(ExprType)

Expression representing a type

type Int = number; // `number` is a type expression

Call(Box<ExprCall>)

Expression to call the lvalue

a(1, 2, 3);